requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
deepface==0.0.85
dm-tree==0.1.8
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
gdown==5.1.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==21.2.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyterlab_widgets==3.0.10
keras==2.15.0
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyfacer @ git+https://github.com/FacePerceiver/facer.git@b2e4fcb94fa8db7cff1a053b6c6441131274c074
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.12.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
validators==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
zipp==3.17.0
